requirements.txt
ray~=2.10.0
numpy~=1.17.4
pillow~=10.2.0
scipy~=1.10.1